created utils for efs
