+++ src/teradata_mcp_server/tools/fs/__init__.py
@@ -1,2 +1,3 @@
 from .fs_resources import *
 from .fs_tools import *
+from .fs_utils import *
